Policy: MOVEONLY: 3 functions to policy
- [script/standard.o] IsStandard
- [main.o] IsStandardTx
- [main.o] AreInputsStandard

coming from btc@9238ecb41752e097443d0bc117df35ebd4ac932e
